Revert "Update MapGeometryModel.cs"
This reverts commit dac2f18196d2f0cc043b5bba8a1dca6e69e4aabb.

+++ Fantome.Libraries.League/IO/MapGeometry/MapGeometryModel.cs
@@ -151,7 +151,7 @@ namespace Fantome.Libraries.League.IO.MapGeometry
 
             this.BoundingBox.Write(bw);
             this.Transformation.Write(bw);
-            bw.Write((byte)this.Flags);
+            bw.Write(this.Flags);
 
             if (version == 7)
             {
